+ разбивка структуры js на модули
